# Perf: avoid concat on flush as well