This only helps when we're not zero padding the last buffered data. Otherwise we'll always have at least 2 buffers and concat.

```js
singleRemainder: test({
	inSize: 150*MB,
	inCount: 1,
	blockSize: 160*MB,
	zeroPadding: false
}),
bigRemainder: test({
	inSize: 200*MB,
	inCount: 1,
	blockSize: 101*MB,
	zeroPadding: false
})
```

Test|Version|AVG
-|-|-
singleRemainder|pull-block master|68ms
singleRemainder|pull-block dev|0ms
bigRemainder|pull-block master|56ms
bigRemainder|pull-block dev|0ms