@Beanow 2017-09-20
Full blown chromium results

Test|Version|Min|AVG|Max
-|-|-:|-:|-:
singleRemainder|master|46ms|59ms|86ms
singleRemainder|dev|0ms|0ms|0ms
bigRemainder|master|30ms|38ms|70ms
bigRemainder|dev|0ms|0ms|0ms
perfectHalves|master|0ms|0ms|1ms
perfectHalves|dev|0ms|0ms|1ms
almostHalves|master|18ms|22ms|30ms
almostHalves|dev|18ms|23ms|53ms
perfect10ths|master|0ms|1ms|2ms
perfect10ths|dev|0ms|1ms|2ms
almost10ths|master|4ms|5ms|8ms
almost10ths|dev|4ms|5ms|8ms
perfect100ths|master|1ms|2ms|3ms
perfect100ths|dev|1ms|2ms|4ms
almost100ths|master|1ms|3ms|6ms
almost100ths|dev|1ms|2ms|3ms
hugeBuffers|master|1ms|1ms|2ms
hugeBuffers|dev|1ms|1ms|2ms
tinyBufs|master|43ms|50ms|81ms
tinyBufs|dev|40ms|50ms|75ms
microBufs|master|51ms|52ms|64ms
microBufs|dev|51ms|59ms|88ms
manyPairs|master|41ms|58ms|87ms
manyPairs|dev|43ms|71ms|79ms
perfectPairs|master|41ms|64ms|76ms
perfectPairs|dev|39ms|60ms|75ms
manyTriplets|master|44ms|72ms|79ms
manyTriplets|dev|44ms|62ms|79ms
bigBuffers|master|1ms|1ms|2ms
bigBuffers|dev|0ms|1ms|4ms