@Beanow 2017-09-21
Fun fact, the original PR #2 was also due to concat, not slice. As a concat of 1 buffer copies the buffer to a new one. This happened a lot before https://github.com/dignifiedquire/pull-block/commit/405d1abc5c5e285b0544e52c0452fade9a503001